Refactor terminal component: extract commands, utils, and components
