+++ src/apps/terminal/commands/clear.ts
@@ -1,0 +1,10 @@
+import { Command } from "../types";
+
+export const clearCommand: Command = {
+  name: "clear",
+  description: "Clear terminal screen",
+  handler: () => ({
+    output: "",
+    isError: false,
+  }),
+};
