+++ src/apps/terminal/commands/pwd.ts
@@ -1,0 +1,10 @@
+import { Command } from "../types";
+
+export const pwdCommand: Command = {
+  name: "pwd",
+  description: "Show current directory",
+  handler: (_, context) => ({
+    output: context.currentPath,
+    isError: false,
+  }),
+};
