+++ src/apps/terminal/commands/whoami.ts
@@ -1,0 +1,10 @@
+import { Command } from "../types";
+
+export const whoamiCommand: Command = {
+  name: "whoami",
+  description: "Display current user",
+  handler: (_, context) => ({
+    output: context.username || "you",
+    isError: false,
+  }),
+};
