+++ src/apps/terminal/components/AnimatedEllipsis.tsx
@@ -1,0 +1,19 @@
+import { useState, useEffect } from "react";
+
+export function AnimatedEllipsis() {
+  const [dots, setDots] = useState("");
+
+  useEffect(() => {
+    const patterns = [".", "..", "...", "..", ".", ".", "..", "..."];
+    let index = 0;
+
+    const interval = setInterval(() => {
+      setDots(patterns[index]);
+      index = (index + 1) % patterns.length;
+    }, 200);
+
+    return () => clearInterval(interval);
+  }, []);
+
+  return <span>{dots}</span>;
+}
